Swtiching weights to double precision as has been already done long ago for all floating point values.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@19638 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/trackFindingCDC/typedefs/include/BasicTypes.h
@@ -36,7 +36,7 @@ namespace Belle2 {
     typedef double FloatType;
 
     /// An additive measure of quality (e.g. logarithms of probabilities)
-    typedef float Weight;
+    typedef double Weight;
 
     /// Constant for the highest possible weight
     extern const Weight HIGHEST_WEIGHT;
